feat: DiaryCoreDataStack Type 구현
